third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_exists-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.mark is defined. 
PASS window.performance.clearMarks is defined. 
PASS window.performance.measure is defined. 
PASS window.performance.clearMeasures is defined. 
Harness: the test ran to completion.
